Support multi-level subcommands
Motivating use case:

  dylan new library --type exe my-lib
  dylan new workspace ...

In programs with a fairly large set of subcommands this just reduces
the top-level clutter and need to stuff everything into a top-level
command by adding hyphens or what-have-you.

By default subcommands support the `--help` option and there is also
a top-level `help` subcommand that accepts any (multi-level) subcommand
for which it displays help. For example,

  dylan new --help
  dylan help new workspace