Better handling of "--"
This is the simplest fix I could see for the second part ("Also, and more
importantly...") of #47 ... simply stuff everything after "--" into a new
`unconsumed-arguments` slot in the parser.

While technically this creates a backward incompatibility, the old code resulted in all
the arguments being added to the final option's value list **in reverse order** and it is
highly unlikely that anyone is depending on that broken behavior.